move SearchBox and Map into app-components
