+++ src/app-components/Reports/index.js
@@ -1,6 +1,6 @@
 import React from "react";
 import TextSection from '../../app-components/TextSection';
-import SearchBox from '../../app-containers/SearchBox';
+import SearchBox from '../../app-components/SearchBox';
 import TabLinks from '../../app-components/TabLinks';
 import { RoutePaths } from "../../app-bundles/routes-bundle";
