+++ src/app-components/SearchBox/SearchBox.test.js
