+++ src/app-pages/Home/index.js
@@ -2,8 +2,8 @@ import React from 'react';
 import Card from '../../app-components/Cards';
 import TextSection from '../../app-components/TextSection';
 import CirclePics from '../../app-components/CirclePics';
-import Map from '../../app-containers/Map';
-import SearchBox from '../../app-containers/SearchBox';
+import Map from '../../app-components/Map';
+import SearchBox from '../../app-components/SearchBox';
 import DropDown from '../../app-components/DropDown';
 import { cardObj, circlePicObj } from './data.js';
 import { connect } from "redux-bundler-react";
